Move classes to their own files.
